requirements.txt
mcp>=1.0.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
